+++ lib/laflem/collections/python/virtualenv/const.py
@@ -1,0 +1,11 @@
+"""
+Constants for the virtualenv collection.
+"""
+PACKAGE_NAME = "laflem"
+
+VIRTUALENV_DEFAULT_FOLDER_NAME = "venv"
+
+VIRTUALENV_FOLDER_NAMES = [
+    VIRTUALENV_DEFAULT_FOLDER_NAME,
+    ".venv",
+]