+++ lib/laflem/collections/self/const.py
@@ -1,0 +1,4 @@
+"""
+Self module constants.
+"""
+PACKAGE_NAME = "laflem"