+++ lib/laflem/exceptions/__init__.py
@@ -1,6 +1,6 @@
-'''
+"""
 Import all exceptions.
-'''
+"""
 
 from .base import *
 from .collections import *