+++ lib/laflem/exceptions/collections.py
@@ -1,10 +1,11 @@
-'''
+"""
 Defined collection exceptions.
-'''
+"""
 
 from .base import FlemException
 
+
 class CollectionNotFound(FlemException):
-  '''
-  Raised when a collection is not found.
-  '''
+    """
+    Raised when a collection is not found.
+    """