+++ lib/laflem/log.py
@@ -1,6 +1,6 @@
-'''
+"""
 This module contains the console log used by laflem.
-'''
+"""
 
 from rich.console import Console